feat: fetch document in JavaScript to support preloading
